Update pgfcircvoltage.tex
COLOR SETTINGS

The behavior of the color option depends on where "color=..." is written (this peculiarity did not change).
(0,0) to[color=red,R,etc.] (2,0): bipole and labels are red.
(0,0) to[R,color=red,etc.] (2,0): bipole only is red.
(0,0) to[color=red,R,color=green,etc.] (2,0): bipole is green and labels are red.
The default color of the arrow is the same as the path color (cf. \draw[color=...]), but it can now be separately set to another color, with the "arrow color" option:
(0,0) to[color=red,R,color=green,arrow color=blue,etc.] (2,0).


PFDCIRCSHAPES.TEX AND PGFCIRCVOLTAGE.TEX
In order to separate the voltage arrow color from the path color, I included the arrows as nodes. I had then to create node shapes for the voltage arrows: "vstraight" for a straight voltage arrow and "vcurved" for the other kind. I had to create an arrow tip shape as well with \pgfdeclarearrow: "voltarr" (it was easy then to have the correct orientation for the arrow tip).