Fix label and annotation positions
The code calls for a rewriting, but for now, this will fix cases
of strange rounding for non-unitary scale factor, and give labels
a bit more breathing space from the components.

One problem was that with the base anchor, the depth of label is
ignored, and so the inner sep has no effect. Using the south anchor
and a suitable inner sep is placing the labels more naturally and
easily.

Fixes issue#91