Partially fix ConTeXt --- added the missing \input
Provide 0.8.3 version for ConTeXt, too. Contributions to add older
version (tested!) are welcome.

I defined \tiny and \small alias for ConTeXt, which are used in the new
components.  Definitely, the project need someone using it in ConTeXt to
help with fixes.

Added a chip to the ConTeXt test file, to catch some error in future,
but still the coverage is at least only partial.