finishing code refactoring
The objective of this code refactoring is basically to make the
management of the to-path commands centralized, and to group
togheter the various definitions of the components.

The is still things to do --- the blocks could be made much more
fine-grained and correctly grouped, but this is the big part of it

vim-style folding markers have been added to most files.