Workaround: add a tunable for source voltage shift
As you can see in issue #747, the `voltage shift` factor is not
always coherent when applied to normal bipoles and sources, especially
if the standard size of the component is different.
Changing the positioning of the voltage elements maintaining backward
compatibility is too difficult for a minor review.
Added a tunable to adjust the amount added for sources.

Fixes #747