Prepare CTAN release 1.8.2 (2025-07-08)
Another significant change for Circuitikz's internals: all the text
anchors of the components are now stable, meaning they are meaningful
after the component has been drawn. That was the case only for part of
the components. It *shouldn't* affect anything for standard usage, but
according to the Ti*k*Z manual, this is "the way".

Additionally, new label positions for transistors, and a much better
alphabetical index in the manual.

- Stabilize all the text anchors (problem [reported by user
  `@JPWiedemann` on GitHub](https://github.com/circuitikz/circuitikz/pull/876), while
  coding the [Circuitikz GUI](https://github.com/Circuit2TikZ/CircuiTikZ-Designer)
- Fix text anchors for blocks `oscillator` and `gridnode`, which were completely bogus
- Add `component text=up` or `...down` for transistor labels
- Documentation fixes for text anchors
- Better index entries (use *seealso*, hyperlink in every entry). Thanks to David Carlisle for the hint.