Preparing CTAN release 1.8.3 (2025-11-23)
The main highlight of this version is the fix of the oo-type sources
and transformers. They used parameters different from width and
height to define the shape, and the anchors were not stable (and
some of them were plain wrong). It also adds several capabilities to
them (anchors, additional windings, and so on). Additionally, IGT
thyristors and a couple of generic shapes (AC/DC symbols) have been
added.

- Add anchors, additional winding, global scale switch and symbol
  rotation to the oo-type component
- Add IGCT thyristors
- Add ac/dc symbols
- Fix definition and stabilize anchors of oo- and ooo-type
  components.
- Fix encoding of the manual, removing some latin1 chars and
  converting to utf8. Why that was working is a mystery.